adding states and overlays
new idea is to mask the us map with individual state maps. the area is known for each state, so with number of pixels and area, we can find out what the 'weight' of each individual state is with respect to each other. once masking is done properly, we can tally individual states and get their contribution to the whole